% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{binarize_mat}
\alias{binarize_mat}
\title{Binarize an interaction matrix}
\usage{
binarize_mat(mat)
}
\arguments{
\item{mat}{Interaction matrix containing raw counts of outcomes of interactions.}
}
\value{
Interaction matrix containing binarized outcomes of interactions.
}
\description{
This function converts raw count data in an interaction matrix to binary outcomes. 
     1 = row beats column more than vice versa; 
     -1 = row loses to column more than vice versa;
     0 = row and column individuals win the same number of times (this could be 0 each)
}
