% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_similarity.R
\name{dyadic_similarity}
\alias{dyadic_similarity}
\title{Compare two rank orderings}
\usage{
dyadic_similarity(order1, order2)
}
\arguments{
\item{order1}{The first rank ordering to be compared. Alternatively, this can
be supplied as an interaction matrix.}

\item{order2}{The second rank ordering to be compared. Alternatively, this can
be supplied as an interaction matrix.}
}
\value{
The proportion of dyadic relationship that are shared by the two orders
        This value is 1 if the orders are identical and 0 if the orders are
        exact opposites.
}
\description{
This function compares two rank orderings. For an order of length \emph{n} 
     there are \emph{n choose 2} dyadic relationships implied by that order. 
     For example, the order {a, b, c} implies that a > b, a  > c, and b > c.
     The dyadic similarity between two orders is the proportion of implied 
     dyadic relationship that are shared by the two orders.
}
