% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{edgelist_to_matrix}
\alias{edgelist_to_matrix}
\title{Convert data in edgelist to interaction matrix}
\usage{
edgelist_to_matrix(edgelist, identities)
}
\arguments{
\item{edgelist}{A 2-column matrix or dataframe with the identities of winners in the first column and losers in the second column.}

\item{identities}{A list of contestant identities (useful if not all contestants are in  the edgelist). 
This list also dictates the order in which contestants are arranged in the resulting matrix.}
}
\value{
Produces an interaction matrix with winners in the rows and losers in the columns. 
        Contestants are arranged according to the order specified by identities.
}
\description{
This function converts data in edgelist format to an interaction matrix
}
