% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_inconsistencies.R
\name{identify_inconsistencies}
\alias{identify_inconsistencies}
\title{Identify inconsistencies in an interaction matrix}
\usage{
identify_inconsistencies(mat)
}
\arguments{
\item{mat}{An interaction matrix with winners as rows and losers as columns. 
The order of the names individuals in the rows and columns 
must be identical and is treated as a rank ordering.}
}
\value{
A list of two elements: \describe{
        \item{is}{a vector of the unique indices of 
        individuals involved in an inconsistency}
        \item{i_count}{is the 
        number of inconsistencies in the matrix}
}
}
\description{
This function identifies inconsistencies in an interaction matrix. Inconsistencies
     are cases where an individual lower in rank beats an individual higher in rank
     more often than vice versa. They are identifiable as the cells below the
     diagonal in a rank-ordered interaciton matrix with values greater than their
     corresponding cell above the diagonal. In a binarized matrix, dyads in an
     inconsistency will have a -1 in their cell above the diagonal and a 1 in 
     their cell below the diagonal.
}
