% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{informed_elo}
\alias{informed_elo}
\title{Elo-rating method informed by prior information}
\usage{
informed_elo(contestants, convention, K = 200, lambda = 100,
  initial.ranks = NULL, interactions)
}
\description{
Elo-rating method informed by prior information
}
