% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Convert data in an interaction matrix to edgelist}
\usage{
get_edgelist(mat)
}
\arguments{
\item{mat}{Interaction matrix containing outcomes of interactions.}
}
\value{
A two-column dataframe with winners in first column and losers in the second column.
}
\description{
This function converts data in an interaction matrix to edgelist format
}
