% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranks.R
\name{plot_ranks}
\alias{plot_ranks}
\title{Plot ranks of individuals in a single social group over multiple study periods}
\usage{
plot_ranks(ranks, ...)
}
\arguments{
\item{ranks}{A dataframe of ranks. There should be one row per contestant per
study period. There should be three columns:
   \describe{
     \item{period}{Study periods. They should appear in chronological order.}
     \item{id}{The identity of each contestant. Each contestant should appear
     once per study period}
     \item{rank}{The rank of each contestant in each study period. This can be
     absolute rank, standardized rank, or elo scores.}
   }}

\item{...}{Arguments to be passed to \code{\link{plot}}}
}
\description{
Plot ranks of individuals in a single social group over multiple study periods
}
